Retricted Pages( hạn chế một số chức năng)

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -24,6 +26,8 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -69,6 +73,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     # Viết phương thức
@@ -81,9 +86,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host: 
+        return HttpResponse('Bạn không được cho phép')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -94,8 +103,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host: 
+        return HttpResponse('Bạn không được cho phép')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
